[#83] If port is occupied, server reports it cannot start
It turns out that the server waited for about 15 seconds before
reporting it didn't start, giving the impression that somehow it started
successfully again. 15 seconds is some historical thing which doesn't
make sense at all: either you can grab the port, or it's not available.
So now it's changed to 0 seconds. Now it works correctly. Test added as
test `tstartagain`.