[#91] All GETs for mime application/json type fail
The previous commit 1c54d52c4c0cbf2857d5b5c3c960255e60397bd0 had a
mistake: The code incorrectly assumed that all calls will have a body if
they are JSON calls. With GET calls, that's obviously incorrect. This
commit fixes that, and updates the `tgetjson` test to test for this.