[#101] $ZC -> $ZCH
This is not a problem in the latest version of YottaDB, but is a problem
in r134 and earlier due to them lacking the code from GT.M V6.3013 which
relaxed the lookup requirements for $Z* characters.

See https://gitlab.com/YottaDB/DB/YDB/-/issues/925 for more information. We
opted to keep the relaxed parsing from GT.M, so currently $ZC is
accepted.