[#92][#104] change read timeout to 1; change nogzip -> gzip;
YDB Web Server will now have a read timeout of 1 second, rather than 10
seconds. The effect of having a read timeout of 10 seconds was that the
child process serving a client will hang around for 10 seconds. For the
YDBGUI testing, that meant that if a process opened a database file, it
will stay open for 10 seconds. This is not desirable, and 1 second
should be plenty for most cases. The previous 10 second timeout is too
excessive, as most end users will not browse to a subsequent page in 10
seconds, and it doesn't take too long to start a new process in case one
is needed.

Change the option of `--nogzip` as it's confusing. Previously, the web
server gzipped by default, and you had to tell it that you don't want to
do that. Now the default is no gzipping is done, and you can do gzipping
by passing a --gzip flag.