[#111] Avoid changing directory in the web server
The server never had a facility to serve static files which are not in
the current directory.  Now, the web server will take a `--directory`
flag to read the static files from that specific directory.

We faced a couple of issues when using the server with the new YDBGUI
which forced a new implementation:

1. Database files using a relative path (in `$ydb_glddir`) disappeared
   because the server changed the directory to
   `$ydb_dist/plugin/etc/ydbgui`.
2. `MUPIP INTEG` requires a writable local directory; and if we switch
   to `$ydb_dist/plugin/etc/ydbgui`, you can't write to that as a normal
   user.

In addition, there is a small change to allow users to combine passing
in an .options array for start-up and also pass command-line flags. It
was previously either/or. Now you can do both. This is to allow the
YDBGUI to set the directory via the .options array and to also allow the
passing of the `--readwrite` flag.