[#113] Options when combined don't work properly
When testing the options together while troubleshooting the GUI, I found
out that an option set such as `--port 8089 --log 3 --debug --directory
/opt/yottadb/r1.36_x86_64/plugin/etc/ydbgui --readwrite` lost the
options for `directory` and `readwrite`.

Debugging the problem showed that the issue was that boolean options
(such as `debug` and `readwrite`) did not eat the space after the
option, resulting in the next option being presented as ` --xxx`.

This resulted in this line quitting early:

```
  380  for  quit:'$$trimleadingstr^%XCMD(.cmdline,"--")  do ; process options
```

This is because, ` --xxx` does not begin with `--`: there is an extra
space.

This is now fixed.

Test `tOptionCombine` tests that a `--directory` entry after a `--gzip`
entry works properly.