[#114] Fix crashes when logging errors in Docker containers as non-root users
If a web server process is running as non-root in a docker container,
when writing an error's output to STDOUT (which is done unconditionally
whether logging is enabled or not), we crashed at `STDOUT+3^%ydbwebreq`
with the following error:

```
%YDB-E-DEVOPENFAIL, Error opening /proc/3481/fd/1
%SYSTEM-E-ENO13, Permission denied
At M source location STDOUT+3^%ydbwebreq
```

As this happens when we are handling an error (and $ECODE is set), we
crash with the Emergency Error Trap (the process terminates). So, as a
result, the original error is never sent back to the client.

The fix is for the parent process to test if STDOUT is writable; and if
not, set a variable which indicates that it's not; if that variable is
set, then STDOUT/STDOUTZW will not attempt to print the error
information out. In addition, the parent process will also print
`Logging will be disabled as /proc/xxx/fd/1 is not writable`.

This is not easily testable in the current testing framework (which is
just an M-Unit test), so no test is provided.

Version updated to 2.0.1.