[#17][#95][#106][#107] Authentication/Authorization Initial Implementation
Introduction
------------
The implementation of Authentication/Authorization will take place over
multiple commits. This commit contains the important pieces, with some
code disabled until the future commits make it possible to use it. This
will be described in detail below.

The implementation is featured enough to enable dependent software to
use it in order not to block future development until it gets all the
features that are needed.

Overall, this commit removes the old HTTP Basic authentication with a
scheme that mirrors how JSON Web Tokens work, except we use opaque
tokens that don't mean anything.

User Documentation
------------------
To specify usernames/passwords/authorization, you need to use either
`--auth-stdin` to be asked interactively about the users or set
environment variable `ydbgui_users` to
`un1:pass1:RO/RW;un2:pass2:RO/RW`. The old `--userpass` has been
removed. If you use `ydbgui_users`, you won't be able to use `:` or `;`
because they are part of the delimiters of the passwords.

To login, POST a JSON of `{ "username": "xxx", "password": "xxx" }` to
`/login`. You will be sent back a token in the body as `{ "token": "xxx"
}`. You will get 401 Unauthorized if username/password is not specified
correctly.

To authenticate each request, send the token in the `Authorization:
Bearer` header. If you don't send it, or send a bad token, you will get
an error of 403 Forbidden.

To logout, POST the token in the body as `{ "token": "xxx" }` to
`/logout`. You will succeed (HTTP 200) whether or not the token is
valid; but you will either get `{ "status": "OK" }` or `{ "status":
"token not found" }` in the body.

If you time out, you will get an HTTP code of 403 Forbidden and a
message of time out.

Right now, both logout and timeout are disabled. See why in the
implementation section.

Token cache is cleaned up every 10 seconds when the server is not
serving requests. This will not work on a continuously busy server; so I
may revisit this implementation in the future.

Implementation
--------------
Usernames and passwords are combined with $ZUT at server start-up time
and are hashed to SINGLE hashes in the parent processes. The user
permission and $ZUT time of start-up of the server are associated with the
hash. The end result is that a single username-password pair gets a
single hash only. This is the hash that is supplied to users as a token
when they log-in.

Since this hash is in the parent process, child processes that logout
cannot delete the hash in the parent process. Therefore, while logout
will work once parent and child processes share data in a future commit,
it doesn't work right now.

Timeout works on the token time (which is the $ZUT time of the start-up
of the server) + 15 minutes. Since this means that all authenticated
logins to the server will fail after 15 minutes in the current
implementation, it's disabled. This will be fixed when we implement
ephemeral tokens.

Same thing for Token cache clean-up (also 15 minutes).

Tests
-----
Test `login` does the following:

- Test protected URL gives you 403
- Login with invalid username/password
- Login with valid username/password and get token
- Test protected URL with token
- Test that `readwrite` flag is tied to the user independent of the server
  start-up flag
- Test protected URL with bad token
- Test simulated timeout code (libcurl-plugin has a memory bug, so I
  wasn't able to finish the test; but I can manually do the tests using
  this command:
```
curl -v -H 'Authorization: Bearer 8kHLl0vaLiiiEuC90v8kJC' localhost:9080/test/simtimeout localhost:9080/test/simtimeout
```
- Logout with valid token
- Logout with invalid token

Misc
----
Fix test tOptionsCombine which did not wait till the server started
before starting to do the tests. This is fudged (like the rest of the
tests that do this) by putting a `hang .1`. This can be done better by
waiting for the port to open, but it too much work for a little benefit,
as I can't really touch the port as that is one of the test cases; so
the only way is to use `ss` or a bash feature to check if the port is
open. For now, let's just hang for .1 seconds.

There is also so minor clean-up of an unused variable in the testing
system.