[Pipeline] Fix failures in tOptionCombine due to binary data
Various pipelines failed due to this error:

```
tOptionCombine^%ydbwebtest - Test combining options (#113) - Error: 150381066,tOptionCombine+18^%ydbwebtest,%YDB-E-BADCHAR, $ZCHAR(176,170,143) is not a valid character in the UTF-8 encoding form
```

This is because the test is structured to use gzip, but the original
test was not written to handle gzipped data. This is now fixed.

My guess is that the original version passed as the actual data in the
gzipped data is not zipped (based on the random number chosen), so it is
always found; but if the number contained repetitions (e.g. 1111), it
may get gzipped, and it won't be found anymore.

The test should pass all the time now.