[Pipeline] Fix curl_easy_perform() can't connect failures
Various tests randomly but rarely have issues with they cannot connect
to the server, even though we checked that the server is running by
connecting to it AFTER we started it.

I thought the main reason for "can't connect" failures was that the
jobbed off server did not start yet. So I added a few lines to detect
that we can connect to a server when it's jobbed off before trying to
use libcurl against it.

However, this turned out not to be enough when various tests are
combined together that use the same port. We found out that a. the
libcurl connection keeps the port open in a CLOSE_WAIT state in order
for it to be re-used and b. it's supposed to disappear with `mupip
stop`, but it wasn't disappearing.

The problem is that we weren't waiting for the server jobs to actually
stop; `mupip stop` is an asynchronous operation, and the processes it's
stopping don't stop immediately at the execution of the command.

The solution is to wait for process to stop by using
`$zgetjpi(job,"ISPROCALIVE")`.

For debugging purposes, as I needed it too many times now, add package
`iproute2` to Docker image as I need `ss` to debug socket issues.