[#119] Libsodium Integration
This commit replaces the previous `mkpasswd` integration to hash
passwords to `libsodium` implemented in the `YDBSodium` plugin
(https://gitlab.com/YottaDB/Util/YDBSodium).

If libsodium is found on the system (via pkg-config), we will install
YDBSodium; otherwise, we will skip its installation and print messages
to the users of the servers that usernames/passwords are not available.

Test `tsodiumerror` verifies that the server recovers gracefully from
libsodium errors.

Misc changes:
- Suppress echo of password
- Remove whois package from Dockerfile, and remove check for mkpasswd in CMake
- Pretty print json file when saving it
- Remove unused testing for ydbgui (which is not correct anymore).