[#127][#128] Clarify database messages; Restore error=""/dev/null"" on job command
Clarify database messages:

- Messages don't say it's a tokens database, confusing users thinking
  it's their own database. Now it clearly says it's a tokens database.

The following two were not fixed, as in the stop context, it's possible
we don't have access to the main processes variables, as the job doing
the stopping is different from the main process. It's not possible to do
the following:
- Deletion message shows even though we are not in --log 1,
  but creation message shows in --log 1.
- Deletion message prints even if there are no users, but a
  database was actually never created in the first place.

error=""/dev/null"" restore:

It was removed in the last commit (53f51ffa) while debugging issues when
re-writing the code but not put back in. Without removing this, if the
server starts in a directory it has no write permissions on, it fails as
it cannot create the mje file.