[#131] Add management of ydb_gbldir env variable
We found out from initial testing that various utilities in the GUI
used `ydb_gbldir` environment variable (e.g. GDE, MUPIP), and so the
server actually needs to set that one too.

Since this is a stateful operation, the tests were enhanced to show
that if we fail after the switch the next call will still get the
default values.

Version bumped to 4.1.1.