[#135] Implement X-YDB-Working-Directory and X-YDB-Env-Vars headers
X-YDB-Working-Directory will specify a working directory for the
specific web service call.

X-YDB-Env-Vars is a semicolon delimited list of a=b series, where a is
the name of the environment variable; and b is its value.

Sample usage:

```
 do &libcurl.init
 do &libcurl.addHeader("X-YDB-Global-Directory: /tmp/testdb.gld")
 do &libcurl.addHeader("X-YDB-Working-Directory: /tmp/")
 do &libcurl.addHeader("X-YDB-Env-Vars: ydb_rel=r200; ydb_dir=/foo/")
 set status=$&libcurl.curl(.httpStatus,.return,"GET","http://127.0.0.1:55728/test/gce")
 do &libcurl.cleanup
```

Tests:

```
/test/cwd -> cwd^%ydbwebapi
/test/env -> env^%ydbwebapi
/test/gce -> gce^%ydbwebapi
```

Also upstream libcurl-wrapper changed the name of its .xc file;
Dockerfile is fixed to match.