[YottaDB/DB/YDB#999][Pipeline] Update tests due to YDB#999
The default `$zgbldir` and environment variable `ydb_gbldir` are both no
longer `$PWD/$ydb_gbldir.gld`. The new default due to YDB#999 is an
empty string.

The empty string in addition caused ^ delimited strings to be
interpreted as a global (e.g. with a^b, if a is empty, ^b is interpreted
as a global in the web server, causing a crash if b is malformed). The
delimiter was changed to a pipe (|).